Use tempfile for test dir, use Path & PathBuf instead of String
Tests in general shouldn't drop content into the current working
directory, we can use `tempfile::tempdir()` to get a temporary directory
to use for the output.

At the same time, adjust `String`s used as paths to be `&Path` or
`PathBuf`.

This provides the env var `REXIT_TEST_OUT_DIR` to allow obtaining the
test output in a non-temporary directory if desired by the developer.